API

API

Functions

ChipSort.bitonic_merge — Method.

bitonic_merge(input_a::Vec{N,T}, input_b::Vec{N,T}) where {N,T}

Merges two SIMD.Vec objects of the same type and size using a bitonic sort network. The inputs are assumed to be sorted. Returns a pair of vectors with the first and second halves of the merged sequence.

source

ChipSort.sort_net — Method.

sort_net(input::Vararg{T, L}) where {L,T}

Applies a sorting network of size L to the input elements, returning a sorted tuple.

The elements must support the min and max functions. In the case of SIMD.Vec objects each "lane" across the vectors will be sorted. Therefore with L vectors of size N this function will produce N sorted sequences of size L, after the data is transposed (see transpose_vecs).

source

ChipSort.transpose_vecs — Method.

transpose_vecs(input::Vararg{Vec{N,T}, L}) where {L,N,T}

Transposes a matrix of L vectors of size N into N vectors of size L. Sizes should be powers of 2.

source

ChipSort.first_stream — Method.

Returns the stream with the smallest first element. If the two streams are empty, returns nothing.

source

Index

ChipSort.bitonic_merge

ChipSort.first_stream

ChipSort.sort_net

ChipSort.transpose_vecs